start updating tests to use already existing weather model data
