+++ .gitignore
@@ -13,7 +13,6 @@ build
 CMake*
 Makefile
 dummy*
-weather_files
 test/data
 *.log
 *.pdf
