+++ test/test_synthetic.py
@@ -1,9 +1,6 @@
-import os, os.path as op
+import os.path as op
 from dataclasses import dataclass
-import subprocess
 from datetime import datetime
-import numpy as np
-import xarray as xr
 
 from RAiDER.llreader import BoundingBox
 from RAiDER.models.weatherModel import make_weather_model_filename
@@ -11,34 +8,7 @@ from RAiDER.losreader import Raytracing, getTopOfAtmosphere
 from RAiDER.utilFcns import lla2ecef, ecef2lla
 from RAiDER.cli.validators import modelName2Module
 
-import pytest
-from test import TEST_DIR
-
-
-def update_yaml(dct_cfg:dict, dst:str='temp.yaml'):
-    """ Write a new yaml file from a dictionary.
-
-    Updates parameters in the default 'raider.yaml' file.
-    Each key:value pair will in 'dct_cfg' will overwrite that in the default
-    """
-    import RAiDER, yaml
-
-    template_file = os.path.join(
-                    os.path.dirname(RAiDER.__file__), 'cli', 'raider.yaml')
-
-    with open(template_file, 'r') as f:
-        try:
-            params = yaml.safe_load(f)
-        except yaml.YAMLError as exc:
-            print(exc)
-            raise ValueError(f'Something is wrong with the yaml file {template_file}')
-
-    params = {**params, **dct_cfg}
-
-    with open(dst, 'w') as fh:
-        yaml.safe_dump(params, fh,  default_flow_style=False)
-
-    return dst
+from test import *
 
 
 def update_model(wm_file:str, wm_eq_type:str, wm_dir:str='weather_files_synth'):
@@ -186,7 +156,9 @@ class StudyArea(object):
         wm_bounds = aoi.calc_buffer_ray(self.los.getSensorDirection(), lookDir=self.los.getLookDirection())
         self.wmObj.set_latlon_bounds(wm_bounds)
         wm_fname  = make_weather_model_filename(self.wmName, self.dt, self.wmObj._ll_bounds)
-        self.path_wm_real = op.join(self.wd, 'weather_files_real', wm_fname)
+        self.path_wm_real = op.join(WM_DIR, wm_fname)
+
+        self.wm_dir_synth = op.join(self.wd, 'weather_files_synth')
 
 
     def setup_region(self):
@@ -276,12 +248,11 @@ def test_hydrostatic_eq(region, mod='ERA-5'):
     ## setup the config files
     SAobj = StudyArea(region, mod)
     dct_cfg      = SAobj.make_config_dict()
-    wm_dir_synth = op.dirname(SAobj.path_wm_real).replace('real', 'synth')
-    dct_cfg['runtime_group']['weather_model_directory'] = wm_dir_synth
+    dct_cfg['runtime_group']['weather_model_directory'] = SAobj.wm_dir_synth
     dct_cfg['download_only'] = False
 
     ## update the weather model; t = p for hydrostatic
-    path_synth = update_model(SAobj.path_wm_real, 'hydro', wm_dir_synth)
+    path_synth = update_model(SAobj.path_wm_real, 'hydro', SAobj.wm_dir_synth)
 
     cfg = update_yaml(dct_cfg)
 
@@ -344,12 +315,11 @@ def test_wet_eq_linear(region, mod='ERA-5'):
     ## setup the config files
     SAobj = StudyArea(region, mod)
     dct_cfg      = SAobj.make_config_dict()
-    wm_dir_synth = op.dirname(SAobj.path_wm_real).replace('real', 'synth')
-    dct_cfg['runtime_group']['weather_model_directory'] = wm_dir_synth
+    dct_cfg['runtime_group']['weather_model_directory'] = SAobj.wm_dir_synth
     dct_cfg['download_only'] = False
 
     ## update the weather model; t = e for wet1
-    path_synth = update_model(SAobj.path_wm_real, 'wet_linear', wm_dir_synth)
+    path_synth = update_model(SAobj.path_wm_real, 'wet_linear', SAobj.wm_dir_synth)
 
     cfg = update_yaml(dct_cfg)
 
@@ -411,12 +381,11 @@ def test_wet_eq_nonlinear(region, mod='ERA-5'):
     ## setup the config files
     SAobj = StudyArea(region, mod)
     dct_cfg      = SAobj.make_config_dict()
-    wm_dir_synth = op.dirname(SAobj.path_wm_real).replace('real', 'synth')
-    dct_cfg['runtime_group']['weather_model_directory'] = wm_dir_synth
+    dct_cfg['runtime_group']['weather_model_directory'] = SAobj.wm_dir_synth
     dct_cfg['download_only'] = False
 
     ## update the weather model; t = e for wet1
-    path_synth = update_model(SAobj.path_wm_real, 'wet_nonlinear', wm_dir_synth)
+    path_synth = update_model(SAobj.path_wm_real, 'wet_nonlinear', SAobj.wm_dir_synth)
 
     cfg = update_yaml(dct_cfg)
