+++ test/weather_files/ERA-5_2020_01_30_T13_52_45_32N_35N_119W_116W.nc
