+++ test/weather_files/ERA-5_2022_08_29_T17_00_01.nc
