+++ test/weather_files/ERA-5_2022_08_29_T17_00_01_69N_72N_159W_153W.nc
